Mac mem info

+++ include/procinfo/procinfo.h
@@ -17,10 +17,10 @@
 
 
 /* Total_Physical_Memory returned in KB */
-int procinfo_total_physical_memory(void);
+long int procinfo_total_physical_memory(void);
 
 /* Total_Virtual_Memory returned in KB */
-int procinfo_total_virtual_memory(void);
+long int procinfo_total_virtual_memory(void);
 
 /* Total Open Files */
-int procinfo_open_files(const char * path);
+long int procinfo_open_files(const char * path);
